pyproject.toml
[project]
name = "video-convert-tools"
version = "0.1.0"
description = "Tools to (batch) convert video files using ffmpeg and NVIDIA codec."
readme = "README.md"
authors = [{ name = "Phil Lies", email = "phil@lies.io" }]
requires-python = ">=3.12"
dependencies = [
    "humanize>=4.14.0",
    "loguru>=0.7.3",
    "rich>=14.2.0",
    "typed-ffmpeg>=3.8.1",
    "typer>=0.20.0",
]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
]

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["mypy>=1.19.0", "pre-commit>=4.5.0", "pytest>=9.0.2", "ruff>=0.14.8"]

[project.scripts]
convert_replace = "video_convert_tools.convert_and_replace:main_cli"
convert_sort = "video_convert_tools.convert_and_sort:main_cli"

[tool.ruff]
line-length = 88
target-version = "py312"
extend-select = [
    "E", # pycodestyle
    "F", # pyflakes
    "W", # pycodestyle
    "B", # flake8 bugbear, common issues like mutable parameter
    "S", # flake8-bandit / security
]
ignore = ["E203", "E501"]

[tool.ruff.lint.extend-per-file-ignores]
# Allow assert, magic values, and imports in functions in test files
"test/**.py" = ["S101", "PLR2004", "PLC0415"]

[tool.mypy]
python_version = "3.12"
strict = true
show_error_codes = true
warn_unused_configs = true
